requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-groq
langserve
python-dotenv
